Added convenience functions rawPool(Color) to CrazyhousePosition.
The only const cast is now encapsulated inside rawPool. It could be eliminated,
but that would require a new class wrapping "const PlayerPool&", so that it can
be returned from rawPool.

Right now this is not possible, because creating a new PlayerPool when it is
not found inside the pool would create either a dangling pointer or a memory
leak. The current solution is to add the PlayerPool directly into the Pool
using map::operator[]. This requires a const_cast, but it should be pretty
safe.