Added createPiece to the KBoard API.
VariantInfo::createPiece is a function which can be used by the interface
(e.g. the ICS module or the EditPosition controller) to add new pieces to a
position.

Its argument is a string description (as suggested by Maurizio) with a meaning
which is strictly defined by the particular variant.

This way one creates a dependency interface -> variant, but it's ok, as long
as the part of the interface which depends on a variant would not have a clear
meaning in general. The ICS module is a clear example of this situation.