Initial support for animations inside the pool.
Pool animations are created in a similar way with respect to board animations,
i.e. using schemes. Another Scheme base class is defined in the Animate::Pool
namespace, and several scheme implementation for pool animations.

A private enqueue function in PiecePool does all the job of converting a scheme
into an animation and running it.