BaseAnimator::warp now works regardless of the current position.
This allows an animator to recover from an error, since the current position
doesn't need to be in sync with the sprite setup anymore for warp to work.
Plus, warp is now called by SimpleAnimator and DropAnimator to ensure that the
final sprite setup is in sync with the final position.