Compile with kdelibs trunk.
Apparently, kdelibs is going to change in source incompatible ways even after
the API freeze. I'm going to tag the alpha release requiring KDE 3.92: that
means that there's no guarantee that it will compile and/or work with newer
KDE snapshots and releases.

I wonder what an API freeze with exceptions is for...