Fixed bug introduced by commit f4afb53.
In some situations sprites where not loaded and failed to appear. This was
because the 'args' parameter was not correctly forwarded to the lua rendering
function by some calls to the loader object (specifically Theme::operator()).

I have now renamed operator() to the more meaningful getPixmap, and replaced
all calls that retrieve a piece sprites with an invocation of pieceSprite.

Convenience functions Board::loadSprite and PiecePool::loadSprite have been
added, that automatically load the sprite with the correct orientation
depending on the flipped state.