[#68] Fix Tori Shogi legalitycheck to detect in-check condition.
This is a hack around a template-instantation problem:
Base::canBeCaptured() instantiates a Base::LegalityCheck<>, so we
cut'n'paste it to be sure.  There may be a cleaner fix, but since this
code will be rewritten for the new component API for 1.0beta, there is
not point in spending more time on it.